+++ ErsatzTV.Core/Scheduling/PlayoutModeSchedulerOne.cs
@@ -47,12 +47,17 @@ public class PlayoutModeSchedulerOne : PlayoutModeSchedulerBase<ProgramScheduleI
                     ? FillerKind.GuideMode
                     : FillerKind.None,
                 CustomTitle = scheduleItem.CustomTitle,
-                WatermarkId = scheduleItem.WatermarkId,
                 PreferredAudioLanguageCode = scheduleItem.PreferredAudioLanguageCode,
                 PreferredAudioTitle = scheduleItem.PreferredAudioTitle,
                 PreferredSubtitleLanguageCode = scheduleItem.PreferredSubtitleLanguageCode,
                 SubtitleMode = scheduleItem.SubtitleMode
             };
+
+            if (scheduleItem.WatermarkId is not null)
+            {
+                playoutItem.Watermarks ??= [];
+                playoutItem.Watermarks.Add(scheduleItem.Watermark);
+            }
 
             List<PlayoutItem> playoutItems = AddFiller(
                 playoutBuilderState,